Fix CI by pinning the version of amqp
Must install the Ubuntu package librabbitmq-dev for the amqp extension
to successfully compile.

Can't install the latest version of the amqp extension as it does not
support the installed version of librabbitmq.

Can't test against PHP 7.0+ as the older version of the amqp extension
can't successfully compile.

Use RabbitMQ 4+ to be compatible with the configurations in
testscenario.

Move rabbitmq and redis service invocations to Travis's "services"
feature.

Fixes #109